Rename jormungandr_cli to jcli
