+++ src/bin/jcli_app/rest/mod.rs
