+++ src/bin/jcli_app/rest/v0/mod.rs
