+++ src/bin/jcli_app/utils/segment_parser.rs
